+++ just-rails-8/setup.sh
@@ -23,7 +23,7 @@ maglev() {
 
   rails new . --database=sqlite3 --skip-action-cable --force
 
-  bundle add maglevcms -v '~> 2.1.0'
+  bundle add maglevcms -v '~> 3.0.0'
   bundle add maglevcms-hyperui-kit -v '~> 2.0.0'
   bundle install
 
@@ -33,12 +33,8 @@ maglev() {
   bundle exec rails g maglev:install
   bundle exec rails g maglev:hyperui:install --force
 
-  # HyperUI Kit 2.0 generates maglev_client_javascript_tags which doesn't
-  # exist in MaglevCMS 2.1.0. Replace with the correct helper.
-  sed -i 's/maglev_client_javascript_tags/maglev_live_preview_client_javascript_tag/' \
-    app/views/theme/layout.html.erb
-
   bundle exec rails maglev:create_site
+  bundle exec rails maglev:publish_site
 }